QATAPP-25890 RHEL dev downstream bug fix
Fix snprintf truncation check in mkPath

In the “pre-flight” check to see if truncation will occur in the call to snprintf() in mkPath() in utils/qzip.c, the conditional fails if MAX_PATH_LEN does not accommodate both path components plus one additional byte; but two bytes are needed, one for the “/” separating the two paths and one for the null terminator.

Instead of “pre-flighting”, detect truncation after the fact; this is potentially faster, and certainly less error-prone.

This is a follow-up to #31, which tried to fix the check but failed to account for the null terminator.

Author: Benjamin A. Beasley <code@musicinmybrain.net>
Author Date: Fri Feb 04 2022 03:13:57 GMT+0800 (中国标准时间)

Change-Id: Icbaf54a0f156986e00a659573f3a592b107f8dca
Signed-off-by: Chengfei Zhu <chengfei.zhu@intel.com>