QATAPP-26207: fix memory leak in doCompressIn
The opData was allocted in stack of doCompressIn, and
as a point parameter of cpaDcCompressData2, when doCompressIn
is completed, opData will be released, however,
dcCompression_ProcessCallback called by doCompressOut still
access opData.

To fixed this issue, I move the opData from stack of doCompressIn
to heap. add opData filed in QzCpaStream_T. This ensure that
opData is not released.

Change-Id: I11ce6b5d2a3706f316b233acc55aeda231dcb397
Signed-off-by: Chengfei Zhu <chengfei.zhu@intel.com>