configure: Include header files depending on method found
When using autoconf to detect QAT library headers, the header
files are found in the qat/ subdirectory of the respective
include/ directory. In other cases with ICP_ROOT, the header
files are directly in the include directory.

Create a C preprocessor define to include the header files
from the correct location. As AC_CONFIG_HEADERS is defined,
the variable is stored in config.h. Since nothing includes
config.h, remove AC_CONFIG_HEADERS from configure.ac to
supply the new define on the command line so that extra
include statements cluttering the files is avoided.

Since the included files are from the library, update their
syntax to #include <file.h>, as the syntax #include "file.h"
refers to the local files. Since the defines are now given
on the command line, add an #ifndef _GNU_SOURCE before
defining that particular variable.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>
Signed-off-by: Chengfei Zhu <chengfei.zhu@intel.com>