test: fix overflow on large files
If the input is a large file and we allocate memory via malloc, we can
cause an overflow when computing the comp_out_sz and decomp_out_sz
(lines 4247 and 4250):

	test_arg[i].comp_out_sz = test_arg[i].src_sz * 2;
	test_arg[i].decomp_out_sz = test_arg[i].src_sz * 5;

This would cause the associated mallocs to fail with ENOMEM.

Fix this by using size_t.

Also fix many other places where the variables from test_arg are copied
into other variables.

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>