QATAPP-29527: [qzip] Fix file permission change after decompression
When extracting the 7z file in qzip, if it is a symbolic link file,
we will modify its file mode like a regular file.
In fact, modifying the mode of the symbolic link file is an error,
which will modify the mode of the linked file, we don't need to
change the mode of symbolic link file.

Change-Id: If627ddbb36f198b60061b48c02c9e98bac3bc112
Signed-off-by: Chengfei Zhu <chengfei.zhu@intel.com>