fixed block_count calculation error
When the qzCompress API encounters a failure in compressing data, it calls qzDeflateStoredBlocks(...) to copy data to the destination buffer along with the Gzip header, block header, and source data. However, there appears to be a calculation error related to the estimation of the output length (out_len) within qzDeflateStoredBlocks(...). This error stems from a miscalculation regarding the block count. In my testing, despite providing src_len as 131328, which is twice the size of the block size (65535), the block count consistently calculates as 1 in the log messages. Ideally, with such a src_len, the block count should be 3. This miscalculation leads to an underestimated out_len by at least 5 bytes per block, resulting in a discrepancy between the final *dest_len and total_out_len in qzCompressCrcExt(...). This discrepancy further causes the "assert(*dest_len == sess->total_out)" in qzCompressCrcExt(...) fail.